Created OSpatialFunctionAbstract class for all spatial Function indexable or not
